Backend/requirements.txt
easyocr==1.7.2
fastapi==0.116.1
langchain_core==0.3.69
langchain_groq==0.3.6
numpy==2.3.1
pandas==2.3.1
pdf2image==1.17.0
Pillow==11.3.0
pydantic==2.11.7
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
SQLAlchemy==2.0.35
uvicorn==0.35.0
